Adding event driven modals to decorateMeta in utils.js (#519)
* Fix race condition with event modals where modal.js contains
the event listener, but consumer events fire before modal.js is loaded.

* Adding test for event based modal in decorateMeta

* Removing failing unit test as it is covered eslewhere

---------
